+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## [Unreleased]
 
+- Support starting plugin even with network issues to download latest server when one is already present. #56
+
 ## 2.6.0
 
 - Support Create workspace edit operation. Fixes #51